move batch data to context, create separate components
