+++ ui/home/Stats.pw.tsx
@@ -5,7 +5,7 @@ import * as blockMock from 'mocks/blocks/block';
 import * as statsMock from 'mocks/stats/index';
 import { test, expect } from 'playwright/lib';
 
-import { HomeBlocksDataContextProvider } from './blocksDataContext';
+import { HomeDataContextProvider } from './homeDataContext';
 import Stats from './Stats';
 
 test.describe('all items', () => {
@@ -19,9 +19,9 @@ test.describe('all items', () => {
     await mockApiResponse('general:stats', statsMock.withBtcLocked);
     await mockApiResponse('general:homepage_blocks', [ blockMock.base, blockMock.base2 ]);
     component = await render(
-      <HomeBlocksDataContextProvider>
+      <HomeDataContextProvider>
         <Stats/>
-      </HomeBlocksDataContextProvider>,
+      </HomeDataContextProvider>,
     );
   });
 
@@ -37,9 +37,9 @@ test('no gas info', async({ render, mockApiResponse, mockEnvs }) => {
   await mockApiResponse('general:stats', statsMock.withoutGasInfo);
   await mockApiResponse('general:homepage_blocks', [ blockMock.base, blockMock.base2 ]);
   const component = await render(
-    <HomeBlocksDataContextProvider>
+    <HomeDataContextProvider>
       <Stats/>
-    </HomeBlocksDataContextProvider>,
+    </HomeDataContextProvider>,
   );
 
   await expect(component).toHaveScreenshot();
@@ -53,9 +53,9 @@ test('4 items default view +@mobile -@default', async({ render, mockApiResponse,
   await mockApiResponse('general:stats', statsMock.base);
   await mockApiResponse('general:homepage_blocks', [ blockMock.base, blockMock.base2 ]);
   const component = await render(
-    <HomeBlocksDataContextProvider>
+    <HomeDataContextProvider>
       <Stats/>
-    </HomeBlocksDataContextProvider>,
+    </HomeDataContextProvider>,
   );
   await expect(component).toHaveScreenshot();
 });
@@ -68,9 +68,9 @@ test('3 items default view +@mobile -@default', async({ render, mockApiResponse,
   await mockApiResponse('general:stats', statsMock.base);
   await mockApiResponse('general:homepage_blocks', [ blockMock.base, blockMock.base2 ]);
   const component = await render(
-    <HomeBlocksDataContextProvider>
+    <HomeDataContextProvider>
       <Stats/>
-    </HomeBlocksDataContextProvider>,
+    </HomeDataContextProvider>,
   );
   await expect(component).toHaveScreenshot();
 });
